Fix menu options on post page
I have updated the links of the sidebar on a post page. The reason is
that the links do not fill their parents so, the user has to click on
the menu text to reach the according section. With this commit, the user
can click beside the text, the action is triggered.

Fixes #24